2.0

+++ src/Form/Field/File.php
@@ -3,31 +3,13 @@
 namespace Encore\Admin\Form\Field;
 
 use Encore\Admin\Form\Field;
-use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Arr;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 
 class File extends Field
 {
     use UploadField;
-
-    /**
-     * Css.
-     *
-     * @var array
-     */
-    protected static $css = [
-        '/vendor/laravel-admin/bootstrap-fileinput/css/fileinput.min.css?v=4.3.7',
-    ];
-
-    /**
-     * Js.
-     *
-     * @var array
-     */
-    protected static $js = [
-        '/vendor/laravel-admin/bootstrap-fileinput/js/plugins/canvas-to-blob.min.js?v=4.3.7',
-        '/vendor/laravel-admin/bootstrap-fileinput/js/fileinput.min.js?v=4.3.7',
-    ];
+    use HasValuePicker;
 
     /**
      * Create a new File instance.
@@ -76,7 +58,7 @@ class File extends Field
         /*
          * Make input data validatable if the column data is `null`.
          */
-        if (array_has($input, $this->column) && is_null(array_get($input, $this->column))) {
+        if (Arr::has($input, $this->column) && is_null(Arr::get($input, $this->column))) {
             $input[$this->column] = '';
         }
 
@@ -89,7 +71,7 @@ class File extends Field
         $rules[$this->column] = $fieldRules;
         $attributes[$this->column] = $this->label;
 
-        return Validator::make($input, $rules, $this->validationMessages, $attributes);
+        return \validator($input, $rules, $this->getValidationMessages(), $attributes);
     }
 
     /**
@@ -101,6 +83,10 @@ class File extends Field
      */
     public function prepare($file)
     {
+        if ($this->picker) {
+            return parent::prepare($file);
+        }
+
         if (request()->has(static::FILE_DELETE_FLAG)) {
             return $this->destroy();
         }
@@ -121,7 +107,13 @@ class File extends Field
     {
         $this->renameIfExists($file);
 
-        $path = $this->storage->putFileAs($this->getDirectory(), $file, $this->name);
+        $path = null;
+
+        if (!is_null($this->storagePermission)) {
+            $path = $this->storage->putFileAs($this->getDirectory(), $file, $this->name, $this->storagePermission);
+        } else {
+            $path = $this->storage->putFileAs($this->getDirectory(), $file, $this->name);
+        }
 
         $this->destroy();
 
@@ -136,6 +128,18 @@ class File extends Field
     protected function preview()
     {
         return $this->objectUrl($this->value);
+    }
+
+    /**
+     * Hides the file preview.
+     *
+     * @return $this
+     */
+    public function hidePreview()
+    {
+        return $this->options([
+            'showPreview' => false,
+        ]);
     }
 
     /**
@@ -155,9 +159,11 @@ class File extends Field
      */
     protected function initialPreviewConfig()
     {
-        return [
-            ['caption' => basename($this->value), 'key' => 0],
-        ];
+        $config = ['caption' => basename($this->value), 'key' => 0];
+
+        $config = array_merge($config, $this->guessPreviewType($this->value));
+
+        return [$config];
     }
 
     /**
@@ -167,23 +173,33 @@ class File extends Field
      */
     public function render()
     {
-        $this->options(['overwriteInitial' => true]);
+        if ($this->picker) {
+            return $this->renderFilePicker();
+        }
+
+        $this->options([
+            'overwriteInitial' => true,
+            'msgPlaceholder'   => trans('admin.choose_file'),
+        ]);
+
         $this->setupDefaultOptions();
 
         if (!empty($this->value)) {
-            $this->attribute('data-initial-preview', filter_var($this->preview(), FILTER_VALIDATE_URL));
+            $this->attribute('data-initial-preview', $this->preview());
             $this->attribute('data-initial-caption', $this->initialCaption($this->value));
 
             $this->setupPreviewOptions();
+            /*
+             * If has original value, means the form is in edit mode,
+             * then remove required rule from rules.
+             */
+            unset($this->attributes['required']);
         }
 
-        $options = json_encode($this->options);
-
-        $this->script = <<<EOT
-
-$("input{$this->getElementClassSelector()}").fileinput({$options});
-
-EOT;
+        $this->addVariables([
+            'options'  => json_encode_options($this->options),
+            'settings' => $this->fileActionSettings,
+        ]);
 
         return parent::render();
     }
