+++ src/Form/Field/UploadField.php
@@ -3,9 +3,9 @@
 namespace Encore\Admin\Form\Field;
 
 use Encore\Admin\Form;
+use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\URL;
-use Illuminate\Support\MessageBag;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 
 trait UploadField
@@ -39,9 +39,60 @@ trait UploadField
     protected $useUniqueName = false;
 
     /**
+     * If use sequence name to store upload file.
+     *
      * @var bool
      */
-    protected $removable = false;
+    protected $useSequenceName = false;
+
+    /**
+     * Retain file when delete record from DB.
+     *
+     * @var bool
+     */
+    protected $retainable = false;
+
+    /**
+     * @var bool
+     */
+    protected $downloadable = false;
+
+    /**
+     * Configuration for setting up file actions for newly selected file thumbnails in the preview window.
+     *
+     * @var array
+     */
+    protected $fileActionSettings = [
+        'showRemove' => false,
+        'showDrag'   => false,
+    ];
+
+    /**
+     * Controls the storage permission. Could be 'private' or 'public'.
+     *
+     * @var string
+     */
+    protected $storagePermission;
+
+    /**
+     * @var array
+     */
+    protected $fileTypes = [
+        'image'  => '/^(gif|png|jpe?g|svg|webp)$/i',
+        'html'   => '/^(htm|html)$/i',
+        'office' => '/^(docx?|xlsx?|pptx?|pps|potx?)$/i',
+        'gdocs'  => '/^(docx?|xlsx?|pptx?|pps|potx?|rtf|ods|odt|pages|ai|dxf|ttf|tiff?|wmf|e?ps)$/i',
+        'text'   => '/^(txt|md|csv|nfo|ini|json|php|js|css|ts|sql)$/i',
+        'video'  => '/^(og?|mp4|webm|mp?g|mov|3gp)$/i',
+        'audio'  => '/^(og?|mp3|mp?g|wav)$/i',
+        'pdf'    => '/^(pdf)$/i',
+        'flash'  => '/^(swf)$/i',
+    ];
+
+    /**
+     * @var string
+     */
+    protected $pathColumn;
 
     /**
      * Initialize the storage instance.
@@ -60,14 +111,20 @@ trait UploadField
      */
     protected function setupDefaultOptions()
     {
-        $defaultOptions = [
+        $defaults = [
+            'theme'                => 'fas',
             'overwriteInitial'     => false,
             'initialPreviewAsData' => true,
+            'msgPlaceholder'       => trans('admin.choose_file'),
             'browseLabel'          => trans('admin.browse'),
+            'cancelLabel'          => trans('admin.cancel'),
             'showRemove'           => false,
             'showUpload'           => false,
-//            'initialCaption'       => $this->initialCaption($this->value),
-            'deleteExtraData' => [
+            'showCancel'           => false,
+            'dropZoneEnabled'      => false,
+            'browseClass'          => admin_color('btn btn-%s'),
+            'initialPreviewShowDelete' => false,
+            'deleteExtraData'      => [
                 $this->formatName($this->column) => static::FILE_DELETE_FLAG,
                 static::FILE_DELETE_FLAG         => '',
                 '_token'                         => csrf_token(),
@@ -76,10 +133,12 @@ trait UploadField
         ];
 
         if ($this->form instanceof Form) {
-            $defaultOptions['deleteUrl'] = $this->form->resource().'/'.$this->form->model()->getKey();
+            $defaults['deleteUrl'] = $this->form->resource().'/'.$this->form->model()->getKey();
         }
 
-        $this->options($defaultOptions);
+        $defaults = array_merge($defaults, ['fileActionSettings' => $this->fileActionSettings]);
+
+        $this->options($defaults);
     }
 
     /**
@@ -89,14 +148,55 @@ trait UploadField
      */
     protected function setupPreviewOptions()
     {
-        if (!$this->removable) {
-            return;
+        $initialPreviewConfig = $this->initialPreviewConfig();
+
+        $this->options(compact('initialPreviewConfig'));
+    }
+
+    /**
+     * @return array|bool
+     */
+    protected function guessPreviewType($file)
+    {
+        $filetype = 'other';
+        $ext = strtok(strtolower(pathinfo($file, PATHINFO_EXTENSION)), '?');
+
+        foreach ($this->fileTypes as $type => $pattern) {
+            if (preg_match($pattern, $ext) === 1) {
+                $filetype = $type;
+                break;
+            }
         }
 
-        $this->options([
-            //'initialPreview'        => $this->preview(),
-            'initialPreviewConfig' => $this->initialPreviewConfig(),
-        ]);
+        $extra = ['type' => $filetype];
+
+        if ($filetype == 'video') {
+            $extra['filetype'] = "video/{$ext}";
+        }
+
+        if ($filetype == 'audio') {
+            $extra['filetype'] = "audio/{$ext}";
+        }
+
+        if ($this->downloadable) {
+            $extra['downloadUrl'] = $this->objectUrl($file);
+        }
+
+        return $extra;
+    }
+
+    /**
+     * Indicates if the underlying field is downloadable.
+     *
+     * @param bool $downloadable
+     *
+     * @return $this
+     */
+    public function downloadable($downloadable = true)
+    {
+        $this->downloadable = $downloadable;
+
+        return $this;
     }
 
     /**
@@ -106,7 +206,20 @@ trait UploadField
      */
     public function removable()
     {
-        $this->removable = true;
+        $this->fileActionSettings['showRemove'] = true;
+        $this->options['initialPreviewShowDelete'] = true;
+
+        return $this;
+    }
+
+    /**
+     * Indicates if the underlying field is retainable.
+     *
+     * @return $this
+     */
+    public function retainable($retainable = true)
+    {
+        $this->retainable = $retainable;
 
         return $this;
     }
@@ -130,20 +243,26 @@ trait UploadField
      *
      * @param string $disk Disks defined in `config/filesystems.php`.
      *
+     * @throws \Exception
+     *
      * @return $this
      */
     public function disk($disk)
     {
-        if (!array_key_exists($disk, config('filesystems.disks'))) {
-            $error = new MessageBag([
-                'title'   => 'Config error.',
-                'message' => "Disk [$disk] not configured, please add a disk config in `config/filesystems.php`.",
-            ]);
+        try {
+            $this->storage = Storage::disk($disk);
+        } catch (\Exception $exception) {
+            if (!array_key_exists($disk, config('filesystems.disks'))) {
+                admin_error(
+                    'Config error.',
+                    "Disk [$disk] not configured, please add a disk config in `config/filesystems.php`."
+                );
 
-            return session()->flash('error', $error);
+                return $this;
+            }
+
+            throw $exception;
         }
-
-        $this->storage = Storage::disk($disk);
 
         return $this;
     }
@@ -210,6 +329,18 @@ trait UploadField
     }
 
     /**
+     * Use sequence name for store upload file.
+     *
+     * @return $this
+     */
+    public function sequenceName()
+    {
+        $this->useSequenceName = true;
+
+        return $this;
+    }
+
+    /**
      * Get store name of upload file.
      *
      * @param UploadedFile $file
@@ -220,6 +351,10 @@ trait UploadField
     {
         if ($this->useUniqueName) {
             return $this->generateUniqueName($file);
+        }
+
+        if ($this->useSequenceName) {
+            return $this->generateSequenceName($file);
         }
 
         if ($this->name instanceof \Closure) {
@@ -248,6 +383,20 @@ trait UploadField
     }
 
     /**
+     * Set path column in has-many related model.
+     *
+     * @param string $column
+     *
+     * @return $this
+     */
+    public function pathColumn($column = 'path')
+    {
+        $this->pathColumn = $column;
+
+        return $this;
+    }
+
+    /**
      * Upload file and delete original file.
      *
      * @param UploadedFile $file
@@ -257,6 +406,10 @@ trait UploadField
     protected function upload(UploadedFile $file)
     {
         $this->renameIfExists($file);
+
+        if (!is_null($this->storagePermission)) {
+            return $this->storage->putFileAs($this->getDirectory(), $file, $this->name, $this->storagePermission);
+        }
 
         return $this->storage->putFileAs($this->getDirectory(), $file, $this->name);
     }
@@ -284,6 +437,10 @@ trait UploadField
      */
     public function objectUrl($path)
     {
+        if ($this->pathColumn && is_array($path)) {
+            $path = Arr::get($path, $this->pathColumn);
+        }
+
         if (URL::isValidUrl($path)) {
             return $path;
         }
@@ -308,14 +465,58 @@ trait UploadField
     }
 
     /**
+     * Generate a sequence name for uploaded file.
+     *
+     * @param UploadedFile $file
+     *
+     * @return string
+     */
+    protected function generateSequenceName(UploadedFile $file)
+    {
+        $index = 1;
+        $extension = $file->getClientOriginalExtension();
+        $original = $file->getClientOriginalName();
+        $new = sprintf('%s_%s.%s', $original, $index, $extension);
+
+        while ($this->storage->exists("{$this->getDirectory()}/$new")) {
+            $index++;
+            $new = sprintf('%s_%s.%s', $original, $index, $extension);
+        }
+
+        return $new;
+    }
+
+    /**
      * Destroy original files.
      *
      * @return void.
      */
     public function destroy()
     {
+        if ($this->retainable) {
+            return;
+        }
+
+        if (method_exists($this, 'destroyThumbnail')) {
+            $this->destroyThumbnail();
+        }
+
         if ($this->storage->exists($this->original)) {
             $this->storage->delete($this->original);
         }
+    }
+
+    /**
+     * Set file permission when stored into storage.
+     *
+     * @param string $permission
+     *
+     * @return $this
+     */
+    public function storagePermission($permission)
+    {
+        $this->storagePermission = $permission;
+
+        return $this;
     }
 }
